src/test/java/com/github/iut/converter/StatementTagConverterTest.java
package com.github.iut.converter;

import com.github.iut.ext.DocumentExt;
import org.junit.jupiter.api.Test;
import org.w3c.dom.Document;

import static com.github.iut.XmlTestUtils.assertXmlEqual;
import static com.github.iut.XmlTestUtils.loadXML;

class StatementTagConverterTest {
    @Test
    public void convert1() throws Exception {
        final Document document = loadXML("xmls/StatementTagConverterTest/source1.xml");
        final Document ret = new StatementTagConverter().convert(DocumentExt.of(document,"testInput")).getDocument();
        assertXmlEqual(ret, "xmls/StatementTagConverterTest/expect/convert1.xml");
    }

    @Test
    public void convert2() throws Exception {
        final Document document = loadXML("xmls/StatementTagConverterTest/source2.xml");
        final Document ret = new StatementTagConverter().convert(DocumentExt.of(document,"testInput")).getDocument();
        assertXmlEqual(ret, "xmls/StatementTagConverterTest/expect/convert2.xml");
    }

    @Test
    public void convert3() throws Exception {
        final Document document = loadXML("xmls/StatementTagConverterTest/source3.xml");
        final Document ret = new StatementTagConverter().convert(DocumentExt.of(document,"testInput")).getDocument();
        assertXmlEqual(ret, "xmls/StatementTagConverterTest/expect/convert3.xml");
    }

    @Test
    public void convert4() throws Exception {
        final Document document = loadXML("xmls/StatementTagConverterTest/source4.xml");
        final Document ret = new StatementTagConverter().convert(DocumentExt.of(document,"testInput")).getDocument();
        assertXmlEqual(ret, "xmls/StatementTagConverterTest/expect/convert4.xml");
    }
}